Email: Do not delete the message from server if the user set it as "Never"
Checking the pop3 account's delete policy when delete the mail.
If the user set the delete policy as "never", then we needn't to
delete the mail from server, but only delete the mail from local.

CRs-Fixed: 625820

+++ src/com/android/email/service/Pop3Service.java
@@ -348,7 +348,8 @@ public class Pop3Service extends Service {
                     // Delete this on the server
                     Pop3Message popMessage =
                             (Pop3Message)remoteFolder.getMessage(currentMsg.mServerId);
-                    if (popMessage != null) {
+                    if (popMessage != null &&
+                            account.getDeletePolicy() != Account.DELETE_POLICY_NEVER) {
                         remoteFolder.deleteMessage(popMessage);
                     }
                 }
